refac:"search fields" add options from db
